Use `bpf_get_socket_cookie` to get pid in cgroup egress without kernel >6.10
From 6.10 you can directly get the pid fro the egress program type.

To support wide use we have to work around that.

This is done by using `bpf_get_socket_cookie` to set a cookie value on the socket and then store this mapped to the pid.

In the egress we can retrieve the cookie and lookup the pid in the map.

In `go` we can then use the pid to retreive the cmdline via the /proc system filesystem.